Adapt API injection to support dashboard panels, moved to util lib
