+++ lib/dashboard/index.js
@@ -58,7 +58,10 @@ app.get('/panel/:bundle/:panel/', function(req, res, next) {
         return;
     }
 
-    res.sendFile(path.join(bundle.dashboard.dir, panel.file));
+    var fileLocation = path.join(bundle.dashboard.dir, panel.file);
+    var html = require('../util').injectAPI(fileLocation, bundle, true);
+
+    res.send(html);
 });
 
 app.get('/panel/:bundle/components/*', function(req, res, next) {
