+++ lib/graphics.js
@@ -49,34 +49,9 @@ app.get('/graphics/:bundleName*', function(req, res, next) {
     // If it's a HTML file, inject the graphic setup script and serve that
     // otherwise, send the file unmodified
     if (resName.endsWith('.html')) {
-        var file = fs.readFileSync(fileLocation);
-        var $ = cheerio.load(file);
+        var html = require('./util').injectAPI(fileLocation, bundle, false);
 
-        var scripts = [
-            '<script src="/socket.io/socket.io.js"></script>',
-            '<script src="/nodecg-api.js"></script>',
-            '<script>var socket = io.connect("//%s/");',
-            'window.nodecg = new NodeCG(%s, socket);</script>'
-        ];
-
-        if (config.login.enabled) {
-            scripts[2] = '<script>var socket = io.connect("//%s/", { query: "token=" + qs["key"] });';
-            scripts.unshift('<script src="/login/QueryString.js"></script>');
-        }
-
-        scripts = scripts.join('\n');
-
-        var partialBundle = {
-            name: bundle.name,
-            config: bundle.config
-        };
-
-        scripts = util.format(scripts, filteredConfig.baseURL, JSON.stringify(partialBundle));
-
-        var currentHead = $('head').html();
-        $('head').html(scripts + currentHead);
-
-        res.send($.html());
+        res.send(html);
     } else {
         res.sendFile(fileLocation);
     }
