+++ lib/util.js
@@ -1,6 +1,9 @@
 'use strict';
 
+var cheerio = require('cheerio');
+var fs = require('fs');
 var tokens = require('./login/tokens');
+var util = require('util');
 
 // Open the sessions database
 var Datastore = require('nedb');
@@ -10,6 +13,7 @@ var sessionDB = new Datastore({ filename: './db/sessions.db', autoload: true });
 sessionDB.persistence.setAutocompactionInterval(300000);
 
 var config = require('./config').getConfig();
+var filteredConfig = require('./config').getFilteredConfig();
 
 exports.authCheck = function (req, res, next) {
     if (!config.login.enabled) {
@@ -54,6 +58,12 @@ exports.authCheck = function (req, res, next) {
     }
 };
 
+exports.destroySession = function(sid, cb) {
+    sessionDB.remove({ sid: sid }, { multi: false }, function (err) {
+        cb(err);
+    });
+};
+
 exports.findSession = function(params, cb) {
     params = params || {};
     sessionDB.findOne(params, function (err, doc) {
@@ -65,10 +75,45 @@ exports.findSession = function(params, cb) {
     });
 };
 
-exports.destroySession = function(sid, cb) {
-    sessionDB.remove({ sid: sid }, { multi: false }, function (err) {
-        cb(err);
-    });
+exports.injectAPI = function(fileLocation, bundle, isDashboard) {
+    var file = fs.readFileSync(fileLocation);
+    var $ = cheerio.load(file);
+
+    // All injections need nodecg-api, duh
+    var scripts = ['<script src="/nodecg-api.js"></script>'];
+
+    if (isDashboard) {
+        // Dashboard panels only need to reference parent socket
+        scripts.push('<script>window.nodecg = new NodeCG(%s, parent.socket);</script>');
+    } else {
+        // Graphics need to create their own socket
+        scripts.unshift('<script src="/socket.io/socket.io.js"></script>');
+
+        if (config.login.enabled) {
+            scripts.push('<script>var socket = io.connect("//%s/", { query: "token=" + qs["key"] });');
+        } else {
+            scripts.push('<script>var socket = io.connect("//%s/");');
+        }
+
+        scripts[2] = util.format(scripts[2], filteredConfig.baseURL);
+        scripts.unshift('<script src="/login/QueryString.js"></script>');
+
+        scripts.push('window.nodecg = new NodeCG(%s, socket);</script>');
+    }
+
+    scripts = scripts.join('\n');
+
+    var partialBundle = {
+        name: bundle.name,
+        config: bundle.config
+    };
+
+    scripts = util.format(scripts, JSON.stringify(partialBundle));
+
+    var currentHead = $('head').html();
+    $('head').html(scripts + currentHead);
+
+    return $.html();
 };
 
 exports.versionCompare = function (v1, v2, options) {
